Adapted tests Removed JDBC and SpringData tests

+++ src/main/java/org/springframework/samples/petclinic/repository/VisitRepository.java
@@ -52,7 +52,7 @@ public class VisitRepository implements PanacheRepository<Visit> {
     }
 
     @SuppressWarnings("unchecked")
-    public List<Visit> findByPetId(Integer petId) {
+    public List<Visit> findByPetId(long petId) {
         Query query = this.em.createQuery("SELECT v FROM Visit v where v.pet.id= :id");
         query.setParameter("id", petId);
         return query.getResultList();
